requirements.txt
customtkinter
requests
python-whois
